requirements.txt
aiohttp
aiosqlite
discord.py==2.5.2
python-dotenv
gspread
